% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid_points.R
\name{create_grid_points}
\alias{create_grid_points}
\title{create_grid_points}
\usage{
create_grid_points(resolution = 0.1)
}
\arguments{
\item{resolution}{The resolution of the grid points, specifying the distance 
between each point in degrees. Default is 0.1 degrees.}
}
\value{
A data.frame object
}
\description{
This function generates a grid of points within the bounding box of the USA 
map. It first loads USA map data using ggplot2::map_data("usa") and filters 
it to include only the main landmass (group == 1). It then calculates the 
bounding box coordinates of the USA map, creates grid points with a specified 
resolution within this bounding box, and checks if each point falls within 
the USA polygon. Grid points outside the USA polygon are excluded from the 
final output.
}
\examples{
create_grid_points(0.8)
create_grid_points()
create_grid_points(4)

}
