% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_temp_trend.R
\name{estimate_temp_trend}
\alias{estimate_temp_trend}
\title{estimate_temp_trend}
\usage{
estimate_temp_trend(data, station_id)
}
\arguments{
\item{data}{A data frame containing temperature data with columns such as 
WBANNO, LST_DATE, and T_DAILY_AVG.}

\item{station_id}{The ID of the weather station for which to estimate the 
temperature trend.}
}
\value{
A data.frame object
}
\description{
This function estimates the temperature trend over time for a specific 
weather station using linear regression. It filters the input data for the 
specified station ID, converts dates to years, fits a linear model using 
lm(), and extracts the trend coefficient (slope) in degrees Celsius per year 
to represent the temperature trend.
}
\examples{
estimate_temp_trend(merging, "53878")
estimate_temp_trend(merging, "3063")

}
