% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interpolations.R
\name{plot_interpolations}
\alias{plot_interpolations}
\title{plot_interpolations}
\usage{
plot_interpolations(
  interpolated_data,
  title = "Gridded Interpolations of Daily Mean Temperature",
  leg = "Celsius"
)
}
\arguments{
\item{interpolated_data}{A data frame containing interpolated data with 
columns for at least latitude (LATITUDE), longitude (LONGITUDE), and a 
variable of interest (value).}

\item{title}{gives a title to the resulting plot, is "Gridded Interpolations 
of Daily Mean Temperature" by default.}

\item{leg}{gives a name to the legend for the resulting plot, is "Celsius" 
by default.}
}
\value{
A data.frame object
}
\description{
This function plots gridded interpolations of daily mean temperature data on 
a map using ggplot2. It takes a data frame with columns for latitude, 
longitude, and a column of interest (value) as input and generates a tile 
plot on a map.
}
\examples{
plot_interpolations(data)
plot_interpolations(data, title = "This is a title", leg = "Years")

}
