% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.extraction.R
\name{time.extraction}
\alias{time.extraction}
\title{time.extraction}
\usage{
\method{time}{extraction}(data, id = NULL, start = NULL, end = NULL)
}
\arguments{
\item{data}{A data frame containing time-related data with columns such as 
WBANNO and LST_DATE.}

\item{id}{Optional. The ID of the station from which to extract data.}

\item{start}{Optional. The start date in the format "Y-m-d" for filtering 
data within a specific time range.}

\item{end}{Optional. The end date in the format "Y-m-d" for filtering data 
within a specific time range.}
}
\value{
A data.frame object
}
\description{
This function extracts data from a given data frame based on station ID, 
start date, and end date criteria. If no station ID is provided, it returns 
the entire data frame. If a station ID is provided but not found in the data, 
an error is raised. It then filters data based on the specified time range 
(start and end dates).
}
\examples{
time.extraction(merging, "53878")
time.extraction(merging, "53878", start = "2024-02-07")
time.extraction(merging, "53878", end = "2024-02-07")
time.extraction(merging, "53878", start = "2024-02-07", end = "2024-02-27")

}
