% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unique_stations}
\alias{unique_stations}
\title{WBANNO,names,state,longtidue,and latitude of all the stations around USA.}
\format{
a dataframe with 236 rows and 5 columns
\describe{
    \item{WBANNO}{station id}
    \item{station_name}{name of the station}
    \item{state}{state where the station locates}
    \item{LONGITUDE}{longitude of the station}
    \item{LATITUDE}{latitude of the station}
}
}
\usage{
unique_stations
}
\description{
A dataset containing station WBANNO, names, respective state information, longitude, latitude for 236 stations.
}
\keyword{datasets}
