% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_cycle.R
\name{yearly_cycle}
\alias{yearly_cycle}
\title{yearly_cycle}
\usage{
yearly_cycle(data, station_id)
}
\arguments{
\item{data}{A data frame containing temperature data with columns such as 
WBANNO, LST_DATE, and T_DAILY_AVG.}

\item{station_id}{The ID of the weather station for which to estimate the 
yearly temperature cycle.}
}
\value{
A data.frame object
}
\description{
This function calculates the yearly temperature cycle for a specific weather 
station using sinusoidal regression based on daily average temperatures. It 
filters the input data for the specified station ID, converts dates to day 
of the year, fits a sinusoidal regression model using lm, and predicts the 
expected temperatures for each day of the year.
}
\examples{
yearly_cycle(merging, "53878")
yearly_cycle(merging, "3063")

}
